testes quinta

+++ src/controllers/experiment-genotipe.controller.ts
@@ -17,7 +17,6 @@ export class ExperimentGenotipeController {
   private printedHistoryController = new PrintHistoryController();
 
   async getAll(options: any) {
-    console.log('🚀 ~ file: experiment-genotipe.controller.ts:20 ~ ExperimentGenotipeController ~ getAll ~ options', options);
     const parameters: object | any = {};
     let orderBy: object | any;
     parameters.AND = [];
@@ -173,15 +172,9 @@ export class ExperimentGenotipeController {
 
       if (options.filterNtFrom || options.filterNtTo) {
         if (options.filterNtFrom && options.filterNtTo) {
-          parameters.nt = JSON.parse(
-            `{"gte": ${Number(options.filterNtFrom)}, "lte": ${Number(
-              options.filterNtTo,
-            )} }`,
-          );
+          parameters.nt = JSON.parse(`{"gte": ${Number(options.filterNtFrom)}, "lte": ${Number(options.filterNtTo)} }`);
         } else if (options.filterNtFrom) {
-          parameters.nt = JSON.parse(
-            `{"gte": ${Number(options.filterNtFrom)} }`,
-          );
+          parameters.nt = JSON.parse(`{"gte": ${Number(options.filterNtFrom)} }`);
         } else if (options.filterNtTo) {
           parameters.nt = JSON.parse(`{"lte": ${Number(options.filterNtTo)} }`);
         }
@@ -436,11 +429,17 @@ export class ExperimentGenotipeController {
 
   async deleteAll(idExperiment: number) {
     try {
-      const response = await this.ExperimentGenotipeRepository.deleteAll(Number(idExperiment));
-      if (response) {
-        return { status: 200, message: 'Parcelas excluídos' };
+      const parcelsToDelete: any = await this.getAll({ idExperiment });
+      const idList = parcelsToDelete[0]?.map(async (parcela: any) => parcela.id);
+      const { status } = await this.printedHistoryController.deleteAll(idList);
+      if (status === 200) {
+        const response = await this.ExperimentGenotipeRepository.deleteAll(Number(idExperiment));
+        if (response) {
+          return { status: 200, message: 'Parcelas excluídos' };
+        }
+        return { status: 400, message: 'Parcelas não excluídos' };
       }
-      return { status: 400, message: 'Parcelas não excluídos' };
+      return { status: 400, message: 'Erro ao excluir parcelas' };
     } catch (error: any) {
       handleError('Parcelas controller', 'DeleteAll', error.message);
       throw new Error('[Controller] - DeleteAll Parcelas erro');
